+++ uspm/parser.h
@@ -10,6 +10,7 @@
 #include <zconf.h>
 #include <string.h>
 #include <stdio.h>
+#include <curl/curl.h>
 
 
 char* concat(const char *s1, const char *s2)
@@ -203,6 +204,37 @@ int remove_from_packages(char *packagename) {
 
     free(out);
 
+    return 0;
+}
+
+size_t write_data(void *ptr, size_t size, size_t nmemb, FILE *stream) {
+    size_t written = fwrite(ptr, size, nmemb, stream);
+    return written;
+}
+
+int download_package(char *mirror, char *package) {
+    CURL *curl;
+    FILE *fp;
+    CURLcode res;
+    char *url = concat(mirror, package);
+    url = concat(url, ".uspm");
+    printf("%s\n", url);
+
+    char outfilename[FILENAME_MAX] = "";
+    strcpy(outfilename, concat(package, ".uspm"));
+    curl = curl_easy_init();
+    if (curl) {
+        fp = fopen(outfilename,"wb");
+        curl_easy_setopt(curl, CURLOPT_URL, url);
+        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_data);
+        curl_easy_setopt(curl, CURLOPT_WRITEDATA, fp);
+        res = curl_easy_perform(curl);
+        /* always cleanup */
+        curl_easy_cleanup(curl);
+        fclose(fp);
+    }
+
+    free(url);
     return 0;
 }
